february_quals/anime_advisor/requirements.txt
aiofiles==0.6.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
fastapi==0.63.0
h11==0.12.0
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
pydantic==1.7.3
PyJWT==2.0.0
python-dotenv==0.15.0
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.3
watchgod==0.6
websockets==8.1
